Refactor store/opentera into a module
