+++ teleop-vue/src/store/index.ts
@@ -3,7 +3,7 @@
 import { createStore } from "vuex";
 
 import Router from "./modules/router";
-import Opentera from "./modules/opentera";
+import { Opentera } from "./modules/opentera";
 
 export default createStore<any>({
   state: {},
