+++ teleop-vue/src/store/modules/opentera/index.ts
@@ -1,0 +1,3 @@
+export { default as Opentera } from "./opentera";
+export * from "./types";
+export * from "./utils";
